+++ SZL_Backend/SZL_Backend/Dto/TagsDto.cs
@@ -2,7 +2,7 @@ namespace SZL_Backend.Dto
 {
     public class TagsDto
     {
-        public int TagId { get; set; }
+        public long TagId { get; set; }
 
         public string? Status { get; set; }
     }
@@ -12,7 +12,7 @@ namespace SZL_Backend.Dto
 {
     public class TagsCreateDto
     {
-        public int? TagId { get; set; }
+        public long? TagId { get; set; }
         public string? Status { get; set; }
     }
 }
